Componentize svg icons and some UI logic
Simplify frontend code. DRY code.
