+++ frontend/src/lib/BookIcon.svelte
@@ -1,0 +1,6 @@
+<svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+  <path
+    d="M6 22C5.45 22 4.97933 21.8043 4.588 21.413C4.196 21.021 4 20.55 4 20V4C4 3.45 4.196 2.979 4.588 2.587C4.97933 2.19567 5.45 2 6 2H18C18.55 2 19.021 2.19567 19.413 2.587C19.8043 2.979 20 3.45 20 4V20C20 20.55 19.8043 21.021 19.413 21.413C19.021 21.8043 18.55 22 18 22H6ZM6 20H18V4H16V10.125C16 10.325 15.9167 10.4707 15.75 10.562C15.5833 10.654 15.4167 10.65 15.25 10.55L13.5 9.5L11.75 10.55C11.5833 10.65 11.4167 10.654 11.25 10.562C11.0833 10.4707 11 10.325 11 10.125V4H6V20ZM11 4H16H11ZM6 4H18H6Z"
+    fill="#001533"
+  />
+</svg>
