+++ frontend/src/lib/EditIcon.svelte
@@ -1,0 +1,6 @@
+<svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+  <path
+    d="M3 17.4624V20.5024C3 20.7824 3.22 21.0024 3.5 21.0024H6.54C6.67 21.0024 6.8 20.9524 6.89 20.8524L17.81 9.94244L14.06 6.19244L3.15 17.1024C3.05 17.2024 3 17.3224 3 17.4624ZM20.71 7.04244C20.8027 6.94993 20.8762 6.84004 20.9264 6.71907C20.9766 6.59809 21.0024 6.46841 21.0024 6.33744C21.0024 6.20648 20.9766 6.07679 20.9264 5.95582C20.8762 5.83485 20.8027 5.72496 20.71 5.63244L18.37 3.29244C18.2775 3.19974 18.1676 3.12619 18.0466 3.07601C17.9257 3.02583 17.796 3 17.665 3C17.534 3 17.4043 3.02583 17.2834 3.07601C17.1624 3.12619 17.0525 3.19974 16.96 3.29244L15.13 5.12244L18.88 8.87244L20.71 7.04244Z"
+    fill="#33445C"
+  />
+</svg>
