+++ frontend/src/lib/ErrorAlertIcon.svelte
@@ -1,0 +1,10 @@
+<svg
+  class="m-auto"
+  width="44"
+  height="38"
+  viewBox="0 0 44 38"
+  fill="none"
+  xmlns="http://www.w3.org/2000/svg"
+>
+  <path d="M24 24H20V14H24V24ZM24 32H20V28H24V32ZM0 38H44L22 0L0 38Z" fill="#B85659" />
+</svg>
