+++ frontend/src/lib/LanguageBasket.svelte
@@ -1,0 +1,119 @@
+<script lang="ts">
+  import { goto } from '$app/navigation'
+  import { page } from '$app/stores'
+  import {
+    langCodesStore,
+    langCountStore,
+    gatewayCodeAndNamesStore,
+    heartCodeAndNamesStore
+  } from '$lib/stores/LanguagesStore'
+  import { resourceTypesStore } from '$lib/stores/ResourceTypesStore'
+  import { langRegExp, getCode, getName, getResourceTypeLangCode } from '$lib/utils'
+  import GlobeIcon from '$lib/GlobeIcon.svelte'
+  import EditIcon from '$lib/EditIcon.svelte'
+  import CloseIcon from '$lib/CloseIcon.svelte'
+
+  function uncheckGatewayLanguage(langCodeAndName: string) {
+    $gatewayCodeAndNamesStore = $gatewayCodeAndNamesStore.filter((item) => item != langCodeAndName)
+    $langCodesStore = $langCodesStore.filter((item) => item != getCode(langCodeAndName))
+    $langCountStore = $langCodesStore.length
+    if ($resourceTypesStore.length > 0) {
+      $resourceTypesStore = $resourceTypesStore.filter((item) => {
+        return getCode(item) != getCode(langCodeAndName)
+      })
+    }
+  }
+  function uncheckHeartLanguage(langCodeAndName: string) {
+    $heartCodeAndNamesStore = $heartCodeAndNamesStore.filter((item) => item != langCodeAndName)
+    $langCodesStore = $langCodesStore.filter((item) => item != getCode(langCodeAndName))
+    $langCountStore = $langCodesStore.length
+    if ($resourceTypesStore.length > 0) {
+      $resourceTypesStore = $resourceTypesStore.filter((item) => {
+        return getResourceTypeLangCode(item) != getCode(langCodeAndName)
+      })
+    }
+  }
+</script>
+
+{#if langRegExp.test($page.url.pathname)}
+  <div class="my-2 flex items-center">
+    <GlobeIcon />
+    <h2 class="ml-2 text-xl font-semibold text-[#33445C]">Language</h2>
+  </div>
+{:else}
+  <div class="my-2 flex items-center justify-between">
+    <div class="flex items-center justify-between">
+      <GlobeIcon />
+      <h2 class="ml-2 text-xl font-semibold text-[#33445C]">Language</h2>
+    </div>
+    <button
+      class="flex rounded bg-white px-4 py-2 text-xl text-[#33445c] hover:bg-[#efefef]"
+      on:click={() => goto('/languages')}
+    >
+      <EditIcon />
+      <span class="ml-2"> Edit </span>
+    </button>
+  </div>
+{/if}
+{#if ($gatewayCodeAndNamesStore && $gatewayCodeAndNamesStore.length > 0) || ($heartCodeAndNamesStore && $heartCodeAndNamesStore.length > 0)}
+  {#each $gatewayCodeAndNamesStore as langCodeAndName}
+    {#if langRegExp.test($page.url.pathname)}
+      <div
+        class="mt-2 flex w-full items-center justify-between
+                 rounded-lg bg-white p-4 text-xl text-[#66768B]"
+      >
+        <div>
+          <span>{getName(langCodeAndName)}</span><span class="ml-2"
+            >({getCode(langCodeAndName)})</span
+          >
+        </div>
+        <button on:click={() => uncheckGatewayLanguage(langCodeAndName)}>
+          <CloseIcon />
+        </button>
+      </div>
+    {:else}
+      <div
+        class="mt-2 flex w-full items-center justify-between
+                 rounded-lg bg-white p-4 text-xl text-[#66768B]"
+      >
+        <div>
+          <span>{getName(langCodeAndName)}</span><span class="ml-2"
+            >({getCode(langCodeAndName)})</span
+          >
+        </div>
+      </div>
+    {/if}
+  {/each}
+  {#each $heartCodeAndNamesStore as langCodeAndName}
+    {#if langRegExp.test($page.url.pathname)}
+      <div
+        class="mt-2 flex w-full items-center justify-between
+                 rounded-lg bg-white p-4 text-xl text-[#66768B]"
+      >
+        <div>
+          <span>{getName(langCodeAndName)}</span><span class="ml-2"
+            >({getCode(langCodeAndName)})</span
+          >
+        </div>
+        <button on:click={() => uncheckHeartLanguage(langCodeAndName)}>
+          <CloseIcon />
+        </button>
+      </div>
+    {:else}
+      <div
+        class="mt-2 flex w-full items-center justify-between
+                 rounded-lg bg-white p-4 text-xl text-[#66768B]"
+      >
+        <div>
+          <span>{getName(langCodeAndName)}</span><span class="ml-2"
+            >({getCode(langCodeAndName)})</span
+          >
+        </div>
+      </div>
+    {/if}
+  {/each}
+{:else}
+  <div class="rounded-lg bg-[#e5e8eb] p-6 text-xl text-[#66768b]">
+    Selections will appear here once a language is selected
+  </div>
+{/if}
