+++ frontend/src/routes/passages/settings/GenerateDocument.svelte
@@ -16,6 +16,7 @@
   import type { PassagesDocumentRequest } from '$lib/passages/models/passage'
   import { toSnakeCase } from '$lib/camel-to-snake-case-util'
   import { omitIdFromPassageReferences } from '$lib/passages/utils'
+  import ErrorAlertIcon from '$lib/ErrorAlertIcon.svelte'
 
   let apiRootUrl = env.PUBLIC_BACKEND_API_URL
   let fileServerUrl: string = env.PUBLIC_FILE_SERVER_URL
@@ -138,16 +139,7 @@
 <div class="bg-white pb-4 pt-12">
   {#if $errorStore}
     <div class="bg-white">
-      <svg
-        class="m-auto"
-        width="44"
-        height="38"
-        viewBox="0 0 44 38"
-        fill="none"
-        xmlns="http://www.w3.org/2000/svg"
-      >
-        <path d="M24 24H20V14H24V24ZM24 32H20V28H24V32ZM0 38H44L22 0L0 38Z" fill="#B85659" />
-      </svg>
+      <ErrorAlertIcon />
       <div class="m-auto"><h3 class="text-center text-[#B85659]">Uh Oh...</h3></div>
       <div class="m-auto">
         <p class="text-xl text-[#B3B9C2]">
