+++ frontend/src/routes/settings/GenerateDocument.svelte
@@ -21,6 +21,8 @@
   import { getCode, getResourceTypeLangCode, getResourceTypeCode } from '$lib/utils'
   import LogRocket from 'logrocket'
   import TaskStatus from './TaskStatus.svelte'
+  import ErrorAlertIcon from '$lib/ErrorAlertIcon.svelte'
+  import EyeIcon from '$lib/EyeIcon.svelte'
 
   let apiRootUrl = env.PUBLIC_BACKEND_API_URL
   let fileServerUrl: string = env.PUBLIC_FILE_SERVER_URL
@@ -195,16 +197,7 @@
 <div class="bg-white pb-4 pt-12">
   {#if $errorStore}
     <div class="bg-white">
-      <svg
-        class="m-auto"
-        width="44"
-        height="38"
-        viewBox="0 0 44 38"
-        fill="none"
-        xmlns="http://www.w3.org/2000/svg"
-      >
-        <path d="M24 24H20V14H24V24ZM24 32H20V28H24V32ZM0 38H44L22 0L0 38Z" fill="#B85659" />
-      </svg>
+      <ErrorAlertIcon />
       <div class="m-auto"><h3 class="text-center text-[#B85659]">Uh Oh...</h3></div>
       <div class="m-auto">
         <p class="text-xl text-[#B3B9C2]">
@@ -280,20 +273,7 @@
               class="gray-gradient hover:gray-gradient-hover w-1/2 rounded-md border-2 border-[#e5e8eb] p-4 text-center"
               on:click={() => viewFromUrl(htmlDownloadUrl)}
             >
-              <svg
-                class="m-auto"
-                width="23"
-                height="16"
-                viewBox="0 0 23 16"
-                fill="none"
-                xmlns="http://www.w3.org/2000/svg"
-              >
-                <path
-                  d="M11.5 0.5C6.5 0.5 2.23 3.61 0.5 8C2.23 12.39 6.5 15.5 11.5 15.5C16.5 15.5 20.77 12.39 22.5 8C20.77 3.61 16.5 0.5 11.5 0.5ZM11.5 13C8.74 13 6.5 10.76 6.5 8C6.5 5.24 8.74 3 11.5 3C14.26 3 16.5 5.24 16.5 8C16.5 10.76 14.26 13 11.5 13ZM11.5 5C9.84 5 8.5 6.34 8.5 8C8.5 9.66 9.84 11 11.5 11C13.16 11 14.5 9.66 14.5 8C14.5 6.34 13.16 5 11.5 5Z"
-                  fill="#1A130B"
-                  fill-opacity="0.8"
-                />
-              </svg>
+              <EyeIcon />
               <span class="p-4 text-xl">View HTML Online</span>
             </button>
           </div>
